Biome caching
Implemented biome caching for center chunk biomes when searching/getting neighboring biomes to a chunk.
Might've fixed out of bounds crash for some ceiling/floor/offset configurations in caverns.

+++ src/main/java/com/personthecat/cavegenerator/data/ClusterSettings.java
@@ -101,7 +101,7 @@ public class ClusterSettings {
     }
 
     private static void copyChance(ClusterSettingsBuilder builder, float chance) {
-        builder.chance(chance).selectionThreshold((1.0F - chance) * 92.0F);
+        builder.chance(chance).selectionThreshold((1.0D - (double)chance) * 92.0D);
     }
 
     private static void copyRadius(ClusterSettingsBuilder builder, Range radius) {
